weandb/requirements.txt
bcrypt==3.1.7
beautifulsoup4==4.8.1
boto3==1.10.19
botocore==1.13.19
bs4==0.0.1
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
Django==2.2.7
django-cors-headers==3.1.1
docutils==0.15.2
gunicorn==20.0.0
idna==2.8
mysql-connector-python==8.0.12
jmespath==0.9.4
Pillow==6.2.1
protobuf==3.10.0
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.8.0
pytz==2019.3
requests==2.22.0
s3transfer==0.2.1
selenium==3.141.0
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
urllib3==1.25.6
